Create a pip.operations.prepare
